ticket-24 | moved custom annotation to a separate package
